pyproject.toml
[project]
name = "segment-app-tflite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<=3.11"
dependencies = [
    "matplotlib>=3.9.4",
    "numpy<2",
    "pillow>=11.3.0",
    "tflite-runtime>=2.14.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "ruff>=0.12.5",
    "snowflake>=1.6.0",
]

[tool.ruff]
line-length = 80
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle (inclui espaçamento, espaços desnecessários)
    "F",    # pyflakes
    "I",    # isort (ordem de imports)
    "Q",    # flake8-quotes (aspas)
]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"
multiline-quotes = "double"
docstring-quotes = "double"
avoid-escape = true

# Ordem alfabética de imports
[tool.isort]
profile = "black"
line_length = 80
force_alphabetical_sort_within_sections = true
force_sort_within_sections = true

# Black para garantir compatibilidade de formatação
[tool.black]
line-length = 80
target-version = ["py311"]
